refactor(cli): extract mergeAndParseRefs into a store-free module
Mirror the PR #682 review fix so the branches converge: the ref-merge
helper moves out of runtime.ts into domains/shared/mergeAndParseRefs.ts
(no bootStore/ke transitive imports), an explicitly empty config
packages list now uniformly means not configured (falls back to global
refs merged over defaults), and the merge semantics are pinned by a
colocated unit test. resolveHelpConfig, updateRefs, and the doctor
package check import the light module directly.
